+++ src/utils/trimCodeBlock.ts
@@ -1,0 +1,12 @@
+/**
+ * Function trimCodeBlock will trim starting and ending code block from the string if it is present.
+ *
+ * Note: This is usefull for post-processing of the result of the chat LLM model
+ *       when the model wraps the result in the (markdown) code block.
+ *
+ */
+
+export function trimCodeBlock(value: string): string {
+    // TODO: !!! Implement
+    return value;
+}