+++ src/utils/trimEndOfCodeBlock.ts
@@ -3,7 +3,8 @@ import spaceTrim from 'spacetrim';
 /**
  * Function trimEndOfCodeBlock will remove ending code block from the string if it is present.
  *
- * This is usefull for post-processing of the result of the completion LLM model if you want to start code block in the prompt but you don't want to end it in the result.
+ * Note: This is usefull for post-processing of the result of the completion LLM model
+ *       if you want to start code block in the prompt but you don't want to end it in the result.
  */
 
 export function trimEndOfCodeBlock(value: string): string {