+++ CHANGELOG.md
@@ -16,5 +16,10 @@
 ## 0.23.0 _(2024-01-25)_
 
 -   You are able to send markdown code block in prompts _(without traces of escaping)_
--   Function `trimEndOfCodeBlock` is not working with escaped code blocks JUST with markdown code blocks
+-   Postprocessing function `trimEndOfCodeBlock` is not working with escaped code blocks JUST with markdown code blocks
 -   Rename `extractBlocksFromMarkdown` to `extractAllBlocksFromMarkdown`
+
+
+## 0.24.0 _(2024-01-25)_
+
+-   Add postprocessing function `trimCodeBlock`